+++ tests_lib/c/test_e.c
@@ -22,13 +22,13 @@ int main() {
   err = futhark_entry_main(ctx, &out_fut, xs_fut, is0_fut);
 
 #if defined(FUTHARK_BACKEND_c) || defined(FUTHARK_BACKEND_multicore)
-  assert(err != 0);
+  assert(err == FUTHARK_PROGRAM_ERROR);
   err = futhark_context_sync(ctx);
   assert(err == 0);
 #else
   assert(err == 0);
   err = futhark_context_sync(ctx);
-  assert(err != 0);
+  assert(err == FUTHARK_PROGRAM_ERROR);
 #endif
 
   char *error = futhark_context_get_error(ctx);